Finishes bare local deployable behavior [5/5]
Repo will be made public now, though actually running the thing
locally won't be super easy without know-how.

Still need to allow a client website to interact with the live server to make
the above easier. But in theory, it is now possible to deploy the stack to
your AWS account and run both the client and the server.